Исправил замечания, изменил подход к отступам, вынес в отдельные файлы классы Track, TrackViewHolder и TrackAdapter
