+++ app/src/main/java/com/example/playlistmaker/Track.kt
@@ -1,0 +1,8 @@
+package com.example.playlistmaker
+
+data class Track(
+    val trackName: String,
+    val artistName: String,
+    val trackTime: String,
+    val artworkUrl100: String
+)
