+++ app/src/main/java/com/example/playlistmaker/TrackAdapter.kt
@@ -1,0 +1,22 @@
+package com.example.playlistmaker
+
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+
+class TrackAdapter(private val trackList: List<Track>) : RecyclerView.Adapter<TrackViewHolder>() {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): TrackViewHolder {
+        val itemView =
+            LayoutInflater.from(parent.context).inflate(R.layout.item_track, parent, false)
+        return TrackViewHolder(itemView)
+    }
+
+    override fun onBindViewHolder(holder: TrackViewHolder, position: Int) {
+        val track = trackList[position]
+        holder.bind(track)
+    }
+
+    override fun getItemCount(): Int {
+        return trackList.size
+    }
+}
