+++ app/src/main/java/com/example/playlistmaker/Utils.kt
@@ -1,0 +1,9 @@
+package com.example.playlistmaker
+
+import android.content.Context
+import android.util.TypedValue
+
+fun Context.toPx(dp: Int): Float = TypedValue.applyDimension(
+    TypedValue.COMPLEX_UNIT_DIP,
+    dp.toFloat(),
+    resources.displayMetrics)
