+++ app/src/main/res/drawable/placeholder_music.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="35dp"
+    android:height="35dp"
+    android:viewportWidth="35"
+    android:viewportHeight="35">
+  <path
+      android:pathData="M3.606,7.919C4.919,6.588 6.231,5.387 6.813,5.631C7.75,6.006 6.813,7.563 6.25,8.481C5.781,9.269 0.888,15.775 0.888,20.313C0.888,22.712 1.788,24.7 3.4,25.9C4.806,26.95 6.663,27.269 8.35,26.763C10.356,26.181 12.006,24.138 14.087,21.569C16.356,18.775 19.394,15.119 21.737,15.119C24.794,15.119 24.831,17.013 25.038,18.475C17.95,19.675 14.95,25.356 14.95,28.544C14.95,31.731 17.65,34.338 20.969,34.338C24.025,34.338 29.013,31.844 29.763,22.9H34.375V18.212H29.744C29.462,15.119 27.7,10.337 22.188,10.337C17.969,10.337 14.35,13.919 12.925,15.663C11.837,17.031 9.063,20.313 8.631,20.763C8.163,21.325 7.356,22.337 6.55,22.337C5.706,22.337 5.2,20.781 5.875,18.737C6.531,16.694 8.5,13.375 9.344,12.137C10.806,10 11.781,8.538 11.781,5.988C11.781,1.919 8.706,0.625 7.075,0.625C4.6,0.625 2.444,2.5 1.975,2.969C1.3,3.644 0.738,4.206 0.325,4.713L3.606,7.919ZM21.025,29.781C20.444,29.781 19.638,29.294 19.638,28.431C19.638,27.306 21.006,24.306 25.019,23.256C24.456,28.3 22.337,29.781 21.025,29.781Z"
+      android:fillColor="#AEAFB4"/>
+</vector>
