+++ app/src/main/res/layout/item_track.xml
@@ -9,21 +9,24 @@
     android:foreground="?android:attr/selectableItemBackground"
     android:orientation="horizontal"
     app:cardCornerRadius="@dimen/cardCornerRadius_8dp"
-    app:cardElevation="4dp">
+    app:cardElevation="4dp"
+    android:paddingStart="@dimen/layout_marginStart_13dp"
+    android:paddingEnd="@dimen/layout_marginEnd_12dp"
+    >
 
     <ImageView
         android:id="@+id/songImage"
         android:layout_width="@dimen/layout_wh45dp"
         android:layout_height="@dimen/layout_wh45dp"
-        android:layout_marginStart="@dimen/layout_marginStart_13dp"
-        android:layout_marginTop="@dimen/layout_margin_top_8dp"
+        android:layout_gravity="center_vertical"
         android:scaleType="fitCenter"
         android:src="@drawable/search" />
 
     <LinearLayout
         android:layout_width="0dp"
-        android:layout_height="match_parent"
+        android:layout_height="wrap_content"
         android:layout_weight="1"
+        android:layout_gravity="center_vertical"
         android:orientation="vertical"
         android:paddingLeft="@dimen/layout_margin_top_8dp">
 
@@ -32,7 +35,6 @@
             style="@style/textItemSongTrack"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_marginTop="@dimen/layout_marginTop_14dp"
             android:ellipsize="end"
             android:maxLines="1"
             android:textSize="@dimen/text_size_16sp"
@@ -54,8 +56,7 @@
         style="@style/imageArrowForSong"
         android:layout_width="@dimen/layout_wh24dp"
         android:layout_height="@dimen/layout_wh24dp"
-        android:layout_marginTop="@dimen/layout_marginTop_18dp"
-        android:layout_marginEnd="@dimen/layout_marginEnd_12dp"
+        android:layout_gravity="center_vertical"
         android:src="@drawable/image_arrow_right" />
 
 </LinearLayout>
